% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectForConfounders.R
\name{ChangeMethod}
\alias{ChangeMethod}
\title{Change the confounding correction method}
\usage{
ChangeMethod(Corr.Object, new.method = c("ols", "theory", "huber", "naive",
  "none"))
}
\arguments{
\item{Corr.Object}{Output from the \code{Correction} function.}

\item{new.method}{New method to use. Can be one of c("ols", "theory", "huber", "naive", "none") and the default is "ols".}
}
\value{
A \code{Correction}-type list
}
\description{
This function quickly re-estimates the confounding adjusted regression coefficients and C using a new user-specified method
}
