% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectForConfounders.R
\name{Correction}
\alias{Correction}
\title{Estimate the effects of interest in omic data while accounting for unobserved covariates}
\usage{
Correction(Y, X, ind.cov, r.confound, max.iter.svd = 1, method = c("ols",
  "theory", "huber", "naive", "none"), shrink.Sigma = F, min.true.eig = 0.1,
  svd.method = "slow")
}
\arguments{
\item{Y}{a p x n data matrix, where p = #of units (i.e. genes) and n = #of samples}

\item{X}{an n x d model matrix, where d = total number of observed covariates to include in the model. This includes the covariates of interest (i.e. disease status), as well as nuisance covariates like the intercept, plate number, DNA concentration, etc.}

\item{ind.cov}{a vector indicating the columns of X that are of interest. For example, if the phenotype of interest is in the second column, ind.cov = c(2)}

\item{r.confound}{The number of latent factors to include in the model. This can be estimated using \code{est.conf.num} and must be >= 0.}

\item{max.iter.svd}{The number of SVD iterations. When p is large (i.e. DNA methlation or gene expression), the default of 1 will suffice. For smaller p (i.e. metabolomic data), a value of 3 is better.}

\item{method}{The confounder correction method to use. Both "ols" and "theory" use the bias-corrected method described in \url{https://arxiv.org/abs/1801.00865v2}, where "ols" weights units by their inverse variance. If neither of these methods is used, we recommend using "huber".}

\item{shrink.Sigma}{If \code{TRUE}, residual variances are shrunk using empirical Bayes with an inverse gamma prior. The default of \code{FALSE} is recommended.}

\item{min.true.eig}{Minimum eigenvalue of (n-d)/p * L'L to consider, where L are the p x r.confound confounding effects. Default is \code{0.1}.}

\item{svd.method}{SVD method. If "fast", \code{irlba} is used. The default is "slow".}
}
\value{
A list. \item{B}{A \code{p} x \code{length(ind.cov)} matrix containing the estimated confounder-adjusted effects of interest for each unit (i.e. gene).} \item{Sigma}{A p-vector containing the estimated residual variances} \item{p.values}{A \code{p} x \code{length(ind.cov)} matrix containing the confounder adjusted P values of interest} \item{p.values.ols}{A \code{p} x \code{length(ind.cov)} matrix containing the unadjusted P values. These SHOULD NOT BE USED EXCEPT FOR COMPARISON} \item{p.values.f}{A length p-vector containing the confounder adjusted F-test p-values. These are only reported if \code{length(ind.cov)} > 1.} \item{z.scores}{A \code{p} x \code{length(ind.cov)} matrix of confounder adjusted Z-scores. These should be N(0,1) under the null hypothesis} \item{z.scores.ols}{A \code{p} x \code{length(ind.cov)} matrix of unadjusted Z-scores. These SHOULD NOT BE USED EXCEPT FOR COMPARISON} \item{C}{The estimated \code{n} x \code{r.confound} matrix of unobserved covariates.} \item{L}{A \code{p} x \code{r.confound} matrix of confounding effects} \item{Omega}{A \code{r.confound} x \code{length(ind.cov)} matrix that estimates Omega.ols = (X[,ind.cov]' X[,ind.cov])^{-1}X[,ind.cov]' C} \item{all.Omegas}{A list containing Omega.ols estimates for all methods} \item{Cov.total}{An \code{n} x \code{r.confound + ncol(X)} design matrix [X C] that can be used in downstream analysis} \item{method}{The confounder correction method used} \item{ind.cov}{The columns of X that are of interest}
}
\description{
This function estimates the regression coefficients of interest while accounting for unobserved covariates (i.e. latent factors/confounding factors), like cell composition and technical effects. It assumes the number of latent factors in known.
}
