% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfounderCorrection.R
\name{est.conf.num}
\alias{est.conf.num}
\title{Estimate the number of latent factors}
\usage{
est.conf.num(Y, X, method = c("bcv", "ed"), max.r, nRepeat = 20, ...)
}
\arguments{
\item{Y}{a p x n data matrix, where p = #of units (i.e. genes) and n = #of samples}

\item{X}{an n x d model matrix, where d = total number of observed covariates to include in the model. This includes the covariates of interest (i.e. disease status), as well as nuisance covariates like the intercept, plate number, DNA concentration, etc.}

\item{method}{Either bi-cross validation ("bcv") or eigenvalue distance ("ed"), with "bcv" being the default. See \url{https://cran.r-project.org/web/packages/cate/cate.pdf} for more details.}

\item{max.r}{Maximum number of latent factors to consider}

\item{nRepeat}{Number of times to perform bcv}

\item{...}{Additional parameters to include in \code{cate::est.factor.num}.}
}
\value{
A \code{cate::est.factor.num} object.
}
\description{
Uses the \code{est.factor.num} function from the \code{CATE} package to estimate the number of latent factors
}
